+++ src/app/item/item.component.ts
@@ -1,5 +1,6 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { AuthService } from '../login/auth.service';
+import { deleteSingle } from '../shared/animations';
 import { FoodDataService } from '../shared/food-data-service.service';
 import { foodObject } from '../shared/IfoodObject';
 import { NotificationsService } from '../shared/notifications-service.service';
@@ -8,11 +9,13 @@ import { NotificationsService } from '../shared/notifications-service.service';
   selector: 'app-item',
   templateUrl: './item.component.html',
   styleUrls: ['./item.component.css'],
-  animations: [],
+  animations: [deleteSingle],
 })
-export class ItemComponent implements OnInit {
+export class ItemComponent {
   @Input() item!: foodObject;
   @Output() itemDeleted = new EventEmitter();
+  confirmDeleteSingle = 0;
+  @Input() confirmDeleteCount!: number;
 
   constructor(
     private foodService: FoodDataService,
@@ -20,15 +23,17 @@ export class ItemComponent implements OnInit {
     private authService: AuthService
   ) {}
 
-  ngOnInit() {}
-
   onDeleteItem() {
+    this.confirmDeleteSingle += 1;
+    if (this.confirmDeleteSingle < 2) return;
+
     this.authService.itemDeletedID = this.item.id!;
     const url = `https://my-list-a7fb0-default-rtdb.europe-west1.firebasedatabase.app/items`;
     this.foodService.deleteSingleItem(url).subscribe(() => {
       this.foodService.listItems.splice(this.item.itemId, 1);
       this.foodService.setUniqueId();
       this.notification.updateListOfNotifiedItems(this.item.id!);
+      this.confirmDeleteSingle = 0;
       this.itemDeleted.emit();
     });
   }
